Refactor components to extract typeBadges
